First functional version
This is still a beta, but you can clone me and vagrant up me and it
should work.
This has only been tested in a Debian Bookworm baremetal on a
KVM/libvirt environment. It worked out of the box and I don't see why it
shouldn't work in a Virtualbox environment or a Windows host; probably
I'll test it in the future but, at the moment, I've got other
priorities.